requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
arrow==1.3.0
asyncpg==0.29.0
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
frozenlist==1.4.1
greenlet==3.1.0
ics==0.7.2
idna==3.8
loguru==0.7.2
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.34
TatSu==5.12.1
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
tzdata==2024.1
yarl==1.9.11
